Alteracoes

+++ src/main/java/br/ifpe/pp2/models/compra/CompraDAO.java
@@ -2,6 +2,6 @@ package br.ifpe.pp2.models.compra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompraDAO extends JpaRepository<Compra, Integer>{
+public interface CompraDAO extends JpaRepository<Compra, Long>{
 
 }
