+++ src/main/java/br/ifpe/pp2/models/produtos/ProdutosDAO.java
@@ -2,6 +2,6 @@ package br.ifpe.pp2.models.produtos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutosDAO extends JpaRepository<Produtos, Integer>{
+public interface ProdutosDAO extends JpaRepository<Produtos, Long>{
 
 }
